+++ changelog.md
@@ -1,2 +1,2 @@
 ### SuperMartijn642's Config Library 1.1.8
-- Initial release of SuperMartijn642's Config Library for Minecraft 1.20.2
+- Initial release of SuperMartijn642's Config Library for NeoForge 1.20.3 & 1.20.4